libpaths.bb: Edit to add $D for paths on target
The previous libpaths.bb recipe failed to create
symlink for /lib64. There were two learnings from that:

1) $D was missing in the paths (which is required so
that it can run on the build host during do_rootfs).
Hence, the result was that the script ran successfully
on the host because both /lib and /lib64 existed there.

Need to be careful to put $D in the right places since
that path is NOT the target of the symlink here.

2) The symlink must be relative and not absolute, or prelink will
try to follow it to /lib on the build host and fail during do_rootfs
(assuming prelink wasn't disabled, ofcourse).

Change-Id: Ia535ca103346c126dc56fbbe4163c51caaff6bf7
Signed-off-by: Iyer, Naveen <naveen.iyer@intel.com>